added methods for getting portfolioentry
