+++ src/main/java/godevenner/divexbackend/portfolio/service/PortfolioEntryService.java
@@ -1,0 +1,12 @@
+package godevenner.divexbackend.portfolio.service;
+
+
+import godevenner.divexbackend.portfolio.model.PortfolioEntry;
+
+import java.util.List;
+
+public interface PortfolioEntryService {
+
+    List<PortfolioEntry> getPortfolioEntries(Integer portfolioId);
+
+}
