+++ src/main/java/godevenner/divexbackend/portfolio/service/PortfolioEntryServiceImpl.java
@@ -1,0 +1,21 @@
+package godevenner.divexbackend.portfolio.service;
+
+import godevenner.divexbackend.portfolio.model.PortfolioEntry;
+import godevenner.divexbackend.portfolio.repository.PortfolioEntryRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class PortfolioEntryServiceImpl implements PortfolioEntryService {
+
+    private final PortfolioEntryRepository portfolioEntryRepository;
+
+    public List<PortfolioEntry> getPortfolioEntries(Integer userId) {
+        return portfolioEntryRepository.findByPortfolioId(userId);
+    }
+
+
+}
