+++ src/main/java/godevenner/divexbackend/portfolio/service/PortfolioService.java
@@ -1,14 +1,13 @@
-package godevenner.divexbackend.portfolio.model.service;
+package godevenner.divexbackend.portfolio.service;
 
 import godevenner.divexbackend.portfolio.model.Portfolio;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service
 public interface PortfolioService {
 
     List<Portfolio> getPortfolios(Integer userId);
+    Portfolio createPortfolio(Portfolio portfolio);
 
 
 }
